refactor: LobbyProvider 구조 개선 및 파일명 변경
- LobbyProvider를 LobbyPresenter와 LobbyState로 분리
    - LobbyState: `apps/omok/lib/presentation/lobby/lobby_state.dart`
    - LobbyPresenter: `apps/omok/lib/presentation/lobby/lobby_presenter.dart`
- 기존 `lobby_provider.dart` 파일 삭제
- LobbyView에서 변경된 LobbyPresenter 참조하도록 수정
- iOS 프로젝트 파일(`project.pbxproj`) 및 `Podfile.lock` 업데이트
- `Runner.xcworkspace/contents.xcworkspacedata` 업데이트
- `pubspec.yaml`에 `retrofit_generator` 및 `hive_generator` 추가
