+++ apps/omok/lib/presentation/lobby/lobby_presenter.dart
@@ -1,20 +1,10 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'lobby_provider.g.dart';
-part 'lobby_provider.freezed.dart';
+import 'lobby_state.dart';
 
-/// 로비 상태를 나타내는 클래스
-@freezed
-class LobbyState with _$LobbyState {
-  const factory LobbyState({
-    @Default(false) bool isLoading,
-    @Default(0) int playerCount,
-    @Default([]) List<String> roomList,
-  }) = _LobbyState;
-}
+part 'lobby_presenter.g.dart';
 
-/// 로비 상태를 관리하는 Provider
+/// 로비 상태를 관리하는 Presenter
 @riverpod
 class LobbyNotifier extends _$LobbyNotifier {
   @override
@@ -25,11 +15,11 @@ class LobbyNotifier extends _$LobbyNotifier {
   /// 게임 방 생성
   Future<void> createRoom() async {
     state = state.copyWith(isLoading: true);
-    
+
     try {
       // TODO: 실제 방 생성 로직 구현
-      await Future.delayed(const Duration(seconds: 1));
-      
+      await Future<void>.delayed(const Duration(seconds: 1));
+
       final newRoomList = List<String>.from(state.roomList)..add('새로운 방');
       state = state.copyWith(
         isLoading: false,
@@ -44,11 +34,11 @@ class LobbyNotifier extends _$LobbyNotifier {
   /// 게임 방 참가
   Future<void> joinRoom(String roomId) async {
     state = state.copyWith(isLoading: true);
-    
+
     try {
       // TODO: 실제 방 참가 로직 구현
-      await Future.delayed(const Duration(seconds: 1));
-      
+      await Future<void>.delayed(const Duration(seconds: 1));
+
       state = state.copyWith(isLoading: false);
     } catch (e) {
       state = state.copyWith(isLoading: false);
